유저페이지에서 게시물 삭제시 권한체크

+++ src/main/webapp/WEB-INF/jsp/usr/article/detail.jsp
@@ -45,6 +45,8 @@ function Delete__Article__Confirm()
 	const result = confirm('정말로 삭제하시겠습니까?');
 	
 	if(result){
+		alert(${rq.loginedMemberUid});
+		if(${rq.loginedMemberUid == article.uid})
 		location.href='delete?aid=${article.aid}&boardCode=${article.boardCode}';
 	}
 }
